backend/requirements.txt
async-timeout==4.0.2
bcrypt==4.0.1
cachelib==0.9.0
click==8.1.3
Deprecated==1.2.13
dnspython==2.2.1
email-validate==1.1.2
filelock==3.8.0
Flask==2.2.2
Flask-Bcrypt==1.0.1
Flask-Cors==3.0.10
Flask-Session==0.4.0
Flask-Sessions==0.1.5
Flask-SQLAlchemy==3.0.2
greenlet==2.0.1
idna==2.10
importlib-metadata==5.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
packaging==21.3
psycopg2-binary==2.9.5
pyparsing==3.0.9
python-dotenv==0.21.0
redis==4.3.4
six==1.16.0
SQLAlchemy==1.4.43
Werkzeug==2.2.2
wrapt==1.14.1
zipp==3.10.0
